pyproject.toml
[project]
name = "ATTACC"
version = "0.1.0"
description = ""
authors = [
    {name = "Don Sabu",email = "itsdonsabu@gmail.com"},
    {name = "Aneeta Jose",email = "itsaneetajose@gmail.com"},
    {name = "Jacob Philip",email = "jacobauxilium@gmail.com"},
    {name = "Pranav P",email = "pranavpkx@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "torch-geometric (>=2.6.1,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
